% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv2sample-ver003.R
\name{pbc.sample}
\alias{pbc.sample}
\title{Edit pbc data to run sample code}
\usage{
pbc.sample()
}
\description{
Edit pbc data in survival package and make it ready to run the sample code in this manual.
}
\seealso{
\code{pbc} in survival package
}
