% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv2sample-ver003.R
\name{plot.surv2sample}
\alias{plot.surv2sample}
\title{Plot method for surv2sample objects}
\usage{
\method{plot}{surv2sample}(x, measure = NULL, baseline = 0, ...)
}
\arguments{
\item{x}{surv2sample object}

\item{measure}{The type of measure used for the plot. When default(=NULL), plot.survfit() is called and KM plots are given.
When "relative time" is specified, a plot of relative percentiles with corresponidng 0.95 confidence intervals is generatead.}

\item{baseline}{Indicates the baseline group, 0/1. Default is 0.}

\item{...}{For further method}
}
\description{
Creates plots from a surv2sample object.
}
\seealso{
\code{plotCI} in plotrix package
}
