% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv2sample-ver003.R
\docType{package}
\name{surv2sampleComp-package}
\alias{surv2sampleComp-package}
\title{Inference for Model-Free Between-Group Parameters For Censored Survival Data}
\description{
Performs inference of several model-free group contrast measures, which include difference/ratio of cumulative incidence rates
at given time points, quantiles, and restricted mean survival times (RMST).
Two kinds of covariate adjustment procedures (i.e., regression and augmentation) for inference of the metrics based on RMST are also included.
}
\references{
Tian L, Zhao L, Wei LJ. Predicting the restricted mean event time with the subject's baseline covariates in survival analysis. Biostatistics 2014, 15, 222-233.

Zhao L, Tian L, Uno H, Solomon S, Pfeffer M, Schindler J, Wei LJ. Utilizing the integrated difference of two survival functions to quantify the treatment contrast for designing, monitoring, and analyzing a comparative clinical study. Clinical Trials 2012, 9, 570-577.
}
\seealso{
flexsurv plotrix survival
}
\author{
Lu Tian, Hajime Uno, Miki Horiguchi

Maintainer: Miki Horiguchi <horiguchimiki@gmail.com>
}
\keyword{survival}
