\name{surv2sampleComp-package}
\alias{surv2sampleComp-package}
\alias{surv2sampleComp}
\docType{package}
\title{
Inference for model-free between-group parameters for censored survival data.
}
\description{
Performs inference of several model-free group contrast measures, which include difference/ratio of cumulative incidence rates at given time points, quantiles, and restricted mean survival times (RMST). Two kinds of covariate adjustment procedures (i.e., regression and augmentation) for inference of the metrics based on RMST are also included.
}
\details{
\tabular{ll}{
Package: \tab surv2sampleComp\cr
Type: \tab Package\cr
Version: \tab 1.0-3\cr
Date: \tab 2013-12-23\cr
License: \tab GLP-2\cr
}

}
\author{Lu Tian, Hajime Uno 

Maintainer: Hajime Uno <huno@jimmy.harvard.edu>

}

\references{
Tian, Lu; Zhao, Lihui; and Wei, LJ. "On the Restricted Mean Event Time in Survival Analysis." (February 2013). Harvard University Biostatistics Working Paper Series. Working Paper 156. http://biostats.bepress.com/harvardbiostat/paper156}

\keyword{ANCOVA, augmentation, generalized gamma, resampling, restricted mean survival time}

\seealso{flexsurv, survival 
}
